// src/app/model/registro.model.ts
export class Registro {
    name: string ;
    motherLastName:  string ;
    fatherLastName:  string ;
    documentNumber: string ;
    email:  string ;
    birthdayDate: string ;
    documentType:  number ;
    job: number ;
    gender:  number ;
    address: string;
    mobile:string;
    country: number;
    level: number;
    geocode: number;
    
}